Move global security tests into own class
Tests with SpringMvc and JAX-RS even if its not specific
to a technology.
